requirements.txt
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
langsmith>=0.2.0
chromadb>=0.5.0
lxml>=5.0.0
